Moved displaySessionLog to the separated file
